Handle Census 2016 datapack import
All attributes present in the 2016 census are now importable. The
primary change has been to use short headers, as the 2016 datapacks only
use this CSV header form.

Due to the difficulties involved with short headers, the long form has
been left for education attributes that are not available yet for 2016.